requirements.txt
cua-computer[all]
cua-agent[all]
transformers
qwen-vl-utils
Pillow
torch
numpydoc
opencv-python
opencv-python-headless
ctransformers
einops
timm
rapidocr
pyautogui
screeninfo